chore: refactor SurveysTable component, remove unused feature flag
