+++ frontend/src/scenes/surveys/SurveyView.tsx
@@ -58,7 +58,6 @@ export function SurveyView({ id }: { id: string }): JSX.Element {
     const { editingSurvey, updateSurvey, stopSurvey, resumeSurvey, duplicateSurvey, setIsDuplicateToProjectModalOpen } =
         useActions(surveyLogic)
     const { deleteSurvey } = useActions(surveysLogic)
-    const { isOnNewEmptyStateExperiment } = useValues(surveysLogic)
     const { currentOrganization } = useValues(organizationLogic)
 
     const hasMultipleProjects = currentOrganization?.teams && currentOrganization.teams.length > 1
@@ -266,13 +265,11 @@ export function SurveyView({ id }: { id: string }): JSX.Element {
                                       key: 'results',
                                       label: 'Results',
                                   }
-                                : isOnNewEmptyStateExperiment
-                                  ? {
-                                        content: <SurveyResultDemo />,
-                                        key: 'results',
-                                        label: 'Results (Demo)',
-                                    }
-                                  : null,
+                                : {
+                                      content: <SurveyResultDemo />,
+                                      key: 'results',
+                                      label: 'Results (Demo)',
+                                  },
                             {
                                 content: <SurveyOverview onTabChange={setTabKey} />,
                                 key: 'overview',
