+++ frontend/src/scenes/surveys/surveysLogic.tsx
@@ -6,17 +6,11 @@ import { actionToUrl, router, urlToAction } from 'kea-router'
 import { lemonToast } from '@posthog/lemon-ui'
 
 import api, { CountedPaginatedResponse } from 'lib/api'
-import { FEATURE_FLAGS } from 'lib/constants'
-import { FeatureFlagsSet, featureFlagLogic as enabledFlagLogic } from 'lib/logic/featureFlagLogic'
+import { featureFlagLogic as enabledFlagLogic } from 'lib/logic/featureFlagLogic'
 import { ProductIntentContext } from 'lib/utils/product-intents'
 import { Scene } from 'scenes/sceneTypes'
 import { sceneConfigurations } from 'scenes/scenes'
-import {
-    SURVEY_CREATED_SOURCE,
-    SURVEY_EMPTY_STATE_EXPERIMENT_VARIANT,
-    SURVEY_PAGE_SIZE,
-    SurveyTemplate,
-} from 'scenes/surveys/constants'
+import { SURVEY_CREATED_SOURCE, SURVEY_PAGE_SIZE, SurveyTemplate } from 'scenes/surveys/constants'
 import { sanitizeSurvey } from 'scenes/surveys/utils'
 import { teamLogic } from 'scenes/teamLogic'
 import { urls } from 'scenes/urls'
@@ -343,12 +337,6 @@ export const surveysLogic = kea<surveysLogicType>([
         },
     })),
     selectors({
-        isOnNewEmptyStateExperiment: [
-            (s) => [s.enabledFlags],
-            (enabledFlags: FeatureFlagsSet): boolean => {
-                return enabledFlags[FEATURE_FLAGS.SURVEY_EMPTY_STATE_V2] === SURVEY_EMPTY_STATE_EXPERIMENT_VARIANT.TEST
-            },
-        ],
         searchedSurveys: [
             (selectors) => [selectors.data, selectors.searchTerm, selectors.filters],
             (data, searchTerm, filters) => {
